refactor with better naming logic, update tests
