+++ client/apps/settings/client/curations/display/index.jsx
@@ -2,9 +2,9 @@ import { set } from 'lodash'
 import PropTypes from 'prop-types'
 import React from 'react'
 
-import { CanvasContainer } from './components/canvas_container.jsx'
+import { Canvas } from './components/canvas/index.jsx'
 import { Campaign } from './components/campaign.jsx'
-import { Panel } from './components/panel.jsx'
+import { Panel } from './components/panel/index.jsx'
 import DropDownItem from 'client/components/drop_down/index.jsx'
 
 export default class DisplayAdmin extends React.Component {
@@ -80,7 +80,7 @@ export default class DisplayAdmin extends React.Component {
                 <div className='display-admin__section-inner'>
                   <Campaign campaign={campaign} index={index} onChange={this.onChange} />
                   <Panel campaign={campaign} index={index} onChange={this.onChange} />
-                  <CanvasContainer campaign={campaign} index={index} onChange={this.onChange} />
+                  <Canvas campaign={campaign} index={index} onChange={this.onChange} />
                 </div>
               }
             </DropDownItem>
