+++ client/apps/settings/client/curations/display/index.styl
@@ -11,6 +11,7 @@
     top 15px
   &__section-inner,
   &__section--campaign,
+  &__section--canvas,
   &__section--panel
     margin 30px 0
   .rich-text--paragraph, .plain-text
